requirements.txt
# Core dependencies
openai-agents>=0.0.3
pydantic>=2.6.0
python-dotenv>=1.0.0

# Optional: REST API interface
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0

# Utilities
rich>=13.7.0       # Pretty CLI output (color-coded logs)
typer>=0.12.0      # CLI argument parsing
